requirements.txt
Flask==3.1.1
SQLAlchemy==2.0.41
python-dotenv==1.1.0
httpx==0.28.1
beautifulsoup4==4.13.4